+++ start/src/App.vue
@@ -1,30 +1,3 @@
-<script lang="ts" setup>
-  import { reactive } from 'vue'
-  import { FMessage } from 'fighting-design'
+<script lang="ts" setup></script>
 
-  const ruleForm = reactive({
-    account: '',
-    password: ''
-  })
-
-  const handelSubmit = (): void => {
-    FMessage.primary('开始提交表单')
-  }
-</script>
-
-<template>
-  {{ ruleForm }}
-  <f-form label-width="60px" :on-submit="handelSubmit" label-position="left">
-    <f-form-item label="账号">
-      <f-input v-model="ruleForm.account" type="text" placeholder="请输入账号" />
-    </f-form-item>
-
-    <f-form-item label="密码">
-      <f-input v-model="ruleForm.password" type="password" placeholder="请输入密码" />
-    </f-form-item>
-
-    <f-form-item>
-      <f-button type="primary" native-type="submit" block>提交表单</f-button>
-    </f-form-item>
-  </f-form>
-</template>
+<template></template>