+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+**更新**
+
 - `f-skeleton` 组件样式细节优化
 - `f-skeleton` 组件新增 `row-gap` 和 `difference` 配置项
+- `f-form` 组件 `on-submit` 回调参数类型修改，由一个对象类型，改为四个参数类型
+- `f-form-item` 组件新增 `validator` 校验方法
+
+**修复**
+
+- 修复 `f-form` 组件没有传递 rules 和 name 也触发表单校验的问题
 
 ## 0.36.1 (2023-05-08)