+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+## 0.37.0 (2023-05-10)
+
 **更新**
 
 - `f-skeleton` 组件样式细节优化